Move newtons method and interpolators to own folder
