+++ demos/odes/adaptive_time_stepper/basic_ode/demo_adaptive_basic_ode.cpp
@@ -8,7 +8,7 @@
 #include "../../../../src/general/initialise.h"
 
 // An interpolator
-#include "../../../../src/problem/cc_newton_interpolator.h"
+#include "../../../../src/interpolators/cc_newton_interpolator.h"
 
 // The required classes to solve Initial Value Problems (IVP)
 // The factory to create the time stepper (integration method)
