+++ src/CMakeLists.txt
@@ -3,6 +3,8 @@
 ADD_SUBDIRECTORY(general)
 ADD_SUBDIRECTORY(data_structures)
 ADD_SUBDIRECTORY(matrices)
+ADD_SUBDIRECTORY(interpolators)
+ADD_SUBDIRECTORY(newtons_method)
 IF (SCICELLXX_USES_VTK)
    ADD_SUBDIRECTORY(vtk)
 ENDIF (SCICELLXX_USES_VTK)
@@ -10,9 +12,10 @@ ADD_SUBDIRECTORY(linear_solvers)
 ADD_SUBDIRECTORY(equations)
 ADD_SUBDIRECTORY(time_steppers)
 ADD_SUBDIRECTORY(problem)
+
 # Once all subdirectories have compiled their stuff then create a VARIABLE to hold their names
 IF (SCICELLXX_USES_VTK)
-   SET(LIB_src_lib time_stepper_lib equations_lib problem_lib linear_solvers_lib vtk_lib matrices_lib data_structures_lib general_lib numerical_recipes_lib argparse_lib PARENT_SCOPE)
+   SET(LIB_src_lib problem_lib time_stepper_lib newtons_method_lib equations_lib linear_solvers_lib vtk_lib interpolators_lib matrices_lib data_structures_lib general_lib numerical_recipes_lib argparse_lib PARENT_SCOPE)
 ELSE ()
-   SET(LIB_src_lib time_stepper_lib equations_lib problem_lib linear_solvers_lib matrices_lib data_structures_lib general_lib numerical_recipes_lib argparse_lib PARENT_SCOPE)
+   SET(LIB_src_lib problem_lib time_stepper_lib newtons_method_lib equations_lib linear_solvers_lib interpolators_lib matrices_lib data_structures_lib general_lib numerical_recipes_lib argparse_lib PARENT_SCOPE)
 ENDIF (SCICELLXX_USES_VTK)
