+++ src/interpolators/cc_newton_interpolator.cpp
