+++ src/interpolators/cc_newton_interpolator.h
