+++ src/newtons_method/CMakeLists.txt
@@ -1,7 +1,6 @@
 # Add source files in order of dependence, the ones with no dependency
 # first then the others
-#SET(SRC_FILES ac_interpolator.cpp cc_newton_interpolator.cpp cc_newtons_method.tpl.cpp ac_problem.cpp ac_ivp_for_odes.cpp ac_mesh_free_problem.tpl.cpp)
-SET(SRC_FILES ac_problem.cpp ac_ivp_for_odes.cpp)
+SET(SRC_FILES cc_newtons_method.cpp)
 
 # Include current directory to the path
 #INCLUDE_DIRECTORIES(${CMAKE_CURRENT_SOURCE_DIR})
@@ -9,9 +8,9 @@ SET(SRC_FILES ac_problem.cpp ac_ivp_for_odes.cpp)
 # Create a library with the above files based on the requested library
 # version
 IF(${SCICELLXX_LIB_TYPE} STREQUAL STATIC)
- ADD_LIBRARY(problem_lib STATIC ${SRC_FILES})
+ ADD_LIBRARY(newtons_method_lib STATIC ${SRC_FILES})
 ELSEIF(${SCICELLXX_LIB_TYPE} STREQUAL SHARED)
- ADD_LIBRARY(problem_lib SHARED ${SRC_FILES})
+ ADD_LIBRARY(newtons_method_lib SHARED ${SRC_FILES})
 ELSE(${SCICELLXX_LIB_TYPE} STREQUAL STATIC)
- ADD_LIBRARY(problem_lib ${SRC_FILES})
+ ADD_LIBRARY(newtons_method_lib ${SRC_FILES})
 ENDIF(${SCICELLXX_LIB_TYPE} STREQUAL STATIC)
