+++ src/newtons_method/cc_newtons_method.cpp
